Fichas en un 80%

+++ LIBRECEA/Deudas.cs
@@ -8,12 +8,12 @@ namespace LIBRECEA
     class Deudas
     {
         public string MATERIAL { get; set; }
-        public Double MONTO { get; set; }
+        public Decimal MONTO { get; set; }
         public DateTime FECHA { get; set; }
         
         public Deudas() { }
 
-        public Deudas(string pIdMat, Double pMonto, DateTime pFec)
+        public Deudas(string pIdMat, Decimal pMonto, DateTime pFec)
         {
             this.MATERIAL = pIdMat;
             this.MONTO = pMonto;
